chore: test keys server request
