+++ Cargo.toml
@@ -93,9 +93,9 @@ wiremock = "0.5.19"
 itertools = "0.11.0"
 sha3 = "0.10.8"
 validator = { version = "0.16.1", features = ["derive"] }
+k256 = "0.13.1"
 
 [dev-dependencies]
-k256 = "0.13.1"
 test-context = "0.1"
 
 [build-dependencies]
